k8s/data-generator/requirements.txt
cassandra-driver==3.28.0
faker==19.6.2